requirements.txt
Django>=5.0,<5.1
gunicorn
python-decouple
djangorestframework>=3.14,<4.0
psycopg[binary]
dj-database-url
requests
whitenoise
django-allauth[socialaccount]
django-allauth-ui
django-widget-tweaks
slippers
stripe